requirements.txt
flask
mysql-connector-python
matplotlib
numpy
scikit-learn
